+++ implementation/js/src/infrastructure/lib+-/marionette/layout.js
@@ -102,11 +102,11 @@
 
 			this.regions = _.extend({}, this.regions, options.regions);
 			
-			//hornor layout configuration with split plug-in
-			this.listenToOnce(this, 'before:render', function(){
-				if(this.layout)
+			//hornor layout configuration through $.split plug-in
+			if(this.layout)
+				this.listenToOnce(this, 'before:render', function(){
 					$(this).split(_.result(this, 'layout'));
-			});
+				});
 			
 			//find region marks after 1-render
 			this.listenToOnce(this, 'render', function(){