+++ CHANGELOG.md
@@ -39,10 +39,11 @@ Change Log
 35. Merged the Context concept into View when using app.get(); (you can now use view as a context)
 36. Added .callout to theme base;
 37. Added .more() api to view; (need example: infinite scrolling)
+38. Added .layout config to view;
+39. Merged $.hsplit/vsplit to $.split
 
 working on ...
 
-* Adding .layout config to view;
 * Adding .lock() api to view;
 * Adding data-export script to tools (npm run data-export);
 * Adding more unit tests;